+++ sdk/node/CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to the HPKV WebSocket Client SDK will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - Released at 2025-04-18
+### Added
+- On and Off functions to subscribe
+- Optional connection configuration
+
+### Changed
+- Improved resource cleanup with proper await/disconnect/destroy sequence
+- Added request queueing to queue pending requests when connection fails and resend after reconnection
+- Added cleanup task for stale requests
+- getConnectionStatus() improved to return more comprehensive status of the client
+
+### Breaking Changes
+- disconnect() function was converted to async
+
 ## [1.1.0]- Released at 2025-04-11
 ### Changed
 - enhanced the API for subscription to changes
+- return type for getConnectionStatus() method changed
 
 ## [1.0.0] - Released at 2025/04/05